C++ APIs for rns<-->direct conversions
