+++ icicle/include/icicle/math/storage.h
@@ -16,8 +16,7 @@ struct
 #ifdef __CUDA_ARCH__
   __align__(LIMBS_ALIGNMENT(1))
 #endif
-    storage<1>
-{
+    storage<1> {
   static constexpr unsigned LC = 1;
   uint32_t limbs[1];
 };
@@ -28,8 +27,7 @@ struct
 #ifdef __CUDA_ARCH__
   __align__(LIMBS_ALIGNMENT(1))
 #endif
-    storage<3>
-{
+    storage<3> {
   static constexpr unsigned LC = 3;
   uint32_t limbs[3];
 };
@@ -40,8 +38,7 @@ struct
 #ifdef __CUDA_ARCH__
   __align__(LIMBS_ALIGNMENT(5))
 #endif
-    storage<5>
-{
+    storage<5> {
   static constexpr unsigned LC = 5;
   uint32_t limbs[5];
 };
@@ -53,8 +50,7 @@ struct
 #ifdef __CUDA_ARCH__
   __align__(LIMBS_ALIGNMENT(10))
 #endif
-    storage<10>
-{
+    storage<10> {
   static constexpr unsigned LC = 10;
   uint32_t limbs[10];
 };
@@ -66,8 +62,7 @@ struct
 #ifdef __CUDA_ARCH__
   __align__(LIMBS_ALIGNMENT(15))
 #endif
-    storage<15>
-{
+    storage<15> {
   static constexpr unsigned LC = 15;
   uint32_t limbs[15];
 };
@@ -78,8 +73,7 @@ struct
 #ifdef __CUDA_ARCH__
   __align__(LIMBS_ALIGNMENT(LIMBS_COUNT))
 #endif
-    storage
-{
+    storage {
   static_assert(LIMBS_COUNT % 2 == 0, "odd number of limbs is not supported\n");
   static constexpr unsigned LC = LIMBS_COUNT;
   union { // works only with even LIMBS_COUNT
@@ -93,7 +87,6 @@ struct
 #ifdef __CUDA_ARCH__
   __align__(LIMBS_ALIGNMENT(LIMBS_COUNT))
 #endif
-    storage_array
-{
+    storage_array {
   storage<LIMBS_COUNT> storages[OMEGAS_COUNT];
 };
