+++ icicle/include/icicle/rings/id.h
@@ -1,5 +1,5 @@
 #pragma once
 
 // Define the ring IDs
-#define LABRADOR 2001
+#define LABRADOR  2001
 #define GREYHOUND 2002
